Deciding equality/subtyping on raw types will never work, so get rid of it
In particular, it'd be nice to prove a form of an inversion lemma for
the subtyping relation, that is if `Maybe t <: t' ⇒ t <: t'`, but it
doesn't necessarily hold for such unrestricted types: consider
t' = Maybe t. Then `Maybe t <: Maybe t` by reflexivity, but
`t <: Maybe t` doesn't hold. Hence, we need to restrict the RHS.