Get rid of the universe nonsense
It makes the interface and the implementation quite confusing, all while
not really adding any features.

Indeed, universes as defined are basically lists of types and their
associated `PgType` instances. But, unless we intend to have different
universes with different instances for the same type (and we don't!), we
don't need that, and we can just instead explicitly and
straightforwardly require the `PgType` instances where needed.

This gets rid of all the universes machinery, and, most importatly, of
the permeating `u` implicit parameters making instances and functions
types quite clumsy.