Keep the modifiers list in `Signature`, and parameterize by read/write operation
The rest of the modifiers doesn't really differ from the nullability
that much, hence it makes sense to treat them similarly.

Parameterizing the `Signature` by the read/write operations allows
distinguishing things that can be null on write (nullable, `DEFAULT`ed
and `SERIAL`, for example) and the ones that can be null on read
(nullable, but not `DEFAULT`ed nor `SERIAL`), thus enabling more precise
types.